+++ Wallet/Logic/TransferCode/InAppDeliveryPayload.swift
